Remove duplicate Application types
